+++ CHANGELOG.md
@@ -7,5 +7,11 @@ This project currently does NOT adheres to [Semantic Versioning](http://semver.o
 
 ## [Unreleased] - 2022-05
 
+### Changed
+- Annotation source, input file format.
+
+### Added
+- Assembly, annotation release.
+
 ### Fixed
 - Column names in adata.var to match what is expected in gEAR. TODO: if n_vars == 1, ids or symbols, this actualy does make a difference... they should be symbols!